+++ libcnb/examples/example-01-basics/src/main.rs
@@ -1,8 +1,6 @@
-use libcnb::data::build_plan::BuildPlan;
-use libcnb::{
-    cnb_runtime, BuildContext, Buildpack, DetectContext, DetectOutcome, GenericError,
-    GenericMetadata, GenericPlatform,
-};
+use libcnb::build::{BuildContext, BuildResult, BuildResultBuilder};
+use libcnb::detect::{DetectContext, DetectResult, DetectResultBuilder};
+use libcnb::{cnb_runtime, Buildpack, GenericError, GenericMetadata, GenericPlatform};
 
 struct BasicBuildpack;
 impl Buildpack for BasicBuildpack {
@@ -10,13 +8,13 @@ impl Buildpack for BasicBuildpack {
     type Metadata = GenericMetadata;
     type Error = GenericError;
 
-    fn detect(&self, _context: DetectContext<Self>) -> libcnb::Result<DetectOutcome, Self::Error> {
-        Ok(DetectOutcome::Pass(BuildPlan::new()))
+    fn detect(&self, _context: DetectContext<Self>) -> libcnb::Result<DetectResult, Self::Error> {
+        Ok(DetectResultBuilder::pass().build())
     }
 
-    fn build(&self, context: BuildContext<Self>) -> libcnb::Result<(), Self::Error> {
+    fn build(&self, context: BuildContext<Self>) -> libcnb::Result<BuildResult, Self::Error> {
         println!("Build runs on stack {}!", context.stack_id);
-        Ok(())
+        Ok(BuildResultBuilder::new().build())
     }
 }