+++ libcnb/examples/example-02-ruby-sample/src/main.rs
@@ -3,11 +3,14 @@ use std::process::{Command, Stdio};
 
 use crate::layers::bundler::BundlerLayerLifecycle;
 use crate::layers::ruby::RubyLayerLifecycle;
-use libcnb::data::build_plan::BuildPlan;
+use libcnb::build::{BuildContext, BuildResult, BuildResultBuilder};
+use libcnb::data::launch::{Launch, Process};
+use libcnb::detect::{DetectContext, DetectResult, DetectResultBuilder};
 use libcnb::layer_lifecycle::execute_layer_lifecycle;
-use libcnb::{cnb_runtime, BuildContext, DetectContext, DetectOutcome, GenericPlatform};
-use libcnb::{data, Buildpack};
+use libcnb::Buildpack;
+use libcnb::{cnb_runtime, GenericPlatform};
 use serde::Deserialize;
+
 mod layers;
 
 struct RubyBuildpack;
@@ -16,17 +19,17 @@ impl Buildpack for RubyBuildpack {
     type Metadata = RubyBuildpackMetadata;
     type Error = anyhow::Error;
 
-    fn detect(&self, context: DetectContext<Self>) -> libcnb::Result<DetectOutcome, Self::Error> {
-        let outcome = if context.app_dir.join("Gemfile.lock").exists() {
-            DetectOutcome::Pass(BuildPlan::new())
+    fn detect(&self, context: DetectContext<Self>) -> libcnb::Result<DetectResult, Self::Error> {
+        let result = if context.app_dir.join("Gemfile.lock").exists() {
+            DetectResultBuilder::pass().build()
         } else {
-            DetectOutcome::Fail
+            DetectResultBuilder::fail().build()
         };
 
-        Ok(outcome)
+        Ok(result)
     }
 
-    fn build(&self, context: BuildContext<Self>) -> libcnb::Result<(), Self::Error> {
+    fn build(&self, context: BuildContext<Self>) -> libcnb::Result<BuildResult, Self::Error> {
         println!("---> Ruby Buildpack");
         println!("---> Download and extracting Ruby");
 
@@ -36,8 +39,25 @@ impl Buildpack for RubyBuildpack {
         install_bundler(&ruby_env)?;
         execute_layer_lifecycle("bundler", BundlerLayerLifecycle { ruby_env }, &context)?;
 
-        write_launch(&context)?;
-        Ok(())
+        Ok(BuildResultBuilder::new()
+            .launch(
+                Launch::new()
+                    .process(Process::new(
+                        "web",
+                        "bundle",
+                        vec!["exec", "ruby", "app.rb"],
+                        false,
+                        true,
+                    )?)
+                    .process(Process::new(
+                        "worker",
+                        "bundle",
+                        vec!["exec", "ruby", "worker.rb"],
+                        false,
+                        false,
+                    )?),
+            )
+            .build())
     }
 }
 
@@ -64,24 +84,4 @@ fn install_bundler(ruby_env: &HashMap<String, String>) -> anyhow::Result<()> {
     } else {
         Err(anyhow::anyhow!("Could not install bundler"))
     }
-}
-
-fn write_launch<B: Buildpack>(context: &BuildContext<B>) -> anyhow::Result<()> {
-    let mut launch_toml = data::launch::Launch::new();
-    let web =
-        data::launch::Process::new("web", "bundle", vec!["exec", "ruby", "app.rb"], false, true)?;
-
-    let worker = data::launch::Process::new(
-        "worker",
-        "bundle",
-        vec!["exec", "ruby", "worker.rb"],
-        false,
-        false,
-    )?;
-
-    launch_toml.processes.push(web);
-    launch_toml.processes.push(worker);
-
-    context.write_launch(launch_toml)?;
-    Ok(())
 }