+++ libcnb/src/build.rs
@@ -4,6 +4,7 @@ use serde::de::DeserializeOwned;
 use serde::Serialize;
 
 use crate::buildpack::Buildpack;
+use crate::data::store::Store;
 use crate::{
     data::{
         buildpack::BuildpackToml, buildpack_plan::BuildpackPlan, launch::Launch,
@@ -12,7 +13,7 @@ use crate::{
     toml_file::{read_toml_file, write_toml_file, TomlFileError},
 };
 
-/// Context for a buildpack's build phase execution.
+/// Context for the build phase execution.
 pub struct BuildContext<B: Buildpack + ?Sized> {
     pub layers_dir: PathBuf,
     pub app_dir: PathBuf,
@@ -21,6 +22,78 @@ pub struct BuildContext<B: Buildpack + ?Sized> {
     pub platform: B::Platform,
     pub buildpack_plan: BuildpackPlan,
     pub buildpack_descriptor: BuildpackToml<B::Metadata>,
+}
+
+/// Describes the result of the build phase.
+///
+/// In contrast to `DetectResult`, it always signals a successful build. To fail the build phase,
+/// return a failed [`libcnb::Result`] from the build function.
+///
+/// It contains build phase output such as launch and/or store metadata which will be subsequently
+/// handled by libcnb.
+///
+/// To construct values of this type, use a [`BuildResultBuilder`].
+#[derive(Debug)]
+pub struct BuildResult(pub(crate) InnerBuildResult);
+
+#[derive(Debug)]
+pub(crate) enum InnerBuildResult {
+    Pass {
+        launch: Option<Launch>,
+        store: Option<Store>,
+    },
+}
+
+/// Constructs [`BuildResult`] values.
+///
+/// # Examples:
+/// ```
+/// use libcnb::build::BuildResultBuilder;
+/// use libcnb_data::launch::{Launch, Process};
+///
+/// let simple = BuildResultBuilder::new().build();
+///
+/// let with_launch = BuildResultBuilder::new()
+///    .launch(Launch::new().process(Process::new("type", "command", vec!["-v"], false, false).unwrap()))
+///    .build();
+/// ```
+pub struct BuildResultBuilder {
+    launch: Option<Launch>,
+    store: Option<Store>,
+}
+
+impl BuildResultBuilder {
+    pub fn new() -> Self {
+        Self {
+            launch: None,
+            store: None,
+        }
+    }
+}
+
+impl BuildResultBuilder {
+    pub fn build(self) -> BuildResult {
+        BuildResult(InnerBuildResult::Pass {
+            launch: self.launch,
+            store: self.store,
+        })
+    }
+
+    pub fn launch(mut self, launch: Launch) -> Self {
+        self.launch = Some(launch);
+        self
+    }
+
+    pub fn store(mut self, store: Store) -> Self {
+        self.store = Some(store);
+        self
+    }
+}
+
+impl Default for BuildResultBuilder {
+    fn default() -> Self {
+        Self::new()
+    }
 }
 
 impl<B: Buildpack> BuildContext<B> {
@@ -102,9 +175,5 @@ impl<B: Buildpack> BuildContext<B> {
         let layer_path = self.layer_path(&layer_name);
         let content_metadata_path = self.layer_content_metadata_path(&layer_name);
         layer_path.exists() && content_metadata_path.exists()
-    }
-
-    pub fn write_launch(&self, data: Launch) -> Result<(), TomlFileError> {
-        write_toml_file(&data, self.layers_dir.join("launch.toml"))
     }
 }