+++ libcnb/src/buildpack.rs
@@ -1,4 +1,6 @@
-use crate::{BuildContext, DetectContext, DetectOutcome, Platform};
+use crate::build::{BuildContext, BuildResult};
+use crate::detect::{DetectContext, DetectResult};
+use crate::Platform;
 use serde::de::DeserializeOwned;
 use std::fmt::{Debug, Display};
 
@@ -7,9 +9,9 @@ pub trait Buildpack {
     type Metadata: DeserializeOwned;
     type Error: Debug + Display;
 
-    fn detect(&self, context: DetectContext<Self>) -> crate::Result<DetectOutcome, Self::Error>;
+    fn detect(&self, context: DetectContext<Self>) -> crate::Result<DetectResult, Self::Error>;
 
-    fn build(&self, context: BuildContext<Self>) -> crate::Result<(), Self::Error>;
+    fn build(&self, context: BuildContext<Self>) -> crate::Result<BuildResult, Self::Error>;
 
     fn handle_error(&self, error: crate::Error<Self::Error>) -> i32 {
         eprintln!("Unhandled error:");