+++ libcnb/src/lib.rs
@@ -27,15 +27,14 @@
 // https://github.com/Malax/libcnb.rs/issues/64
 #![allow(clippy::unnecessary_wraps)]
 
+pub mod build;
+pub mod detect;
 pub mod layer_env;
-
 pub mod layer_lifecycle;
 
 use crate::data::buildpack::BuildpackApi;
-pub use build::BuildContext;
 pub use buildpack::Buildpack;
-pub use detect::DetectContext;
-pub use detect::DetectOutcome;
+
 pub use env::*;
 pub use error::*;
 pub use generic::*;
@@ -44,9 +43,7 @@ pub use platform::*;
 pub use runtime::cnb_runtime;
 pub use toml_file::*;
 
-mod build;
 mod buildpack;
-mod detect;
 mod env;
 mod error;
 mod generic;